gitignore platform specific cython-generated files
Prevent platform specific cython files from being tracked in repo in the
future by adding them to a gitignore.
The rules should be strict enough to avoid anything that shouldn't be
blocked being blocked, but this also means that if more statically
compiled files are added in the future, they may need to be added to the
rules.