Fixed polar (cartesian) grids
Made polar grids symmetric about their poles and extended them by a single grid point in each direction. This seems to have smoothed out the discontinuity between the polar and the cartesian regimes, but the issue of checkerboarding in atmosp_addition remains.

Also removed hardcoding of resolution and polar boundaries in some of the top_level functions.